Implement post-processing part in dudect
According to the paper, cropping data before t_test is indispensable.
To do that, I retrieve the missing code "percentile" and adjust to
fit the form of dudect in lab0.

To be more specific, the minimum of 100 data in exec_times is disposed.
The exponential part tends to maintain the overall distribution of
exec_times.

Change-Id: I2da2cff6b19af0f83b48d752f2e6e7aff399afa1